Commit #6: Created a BlackBoard. Added a static BlackBoard to the I_State header so it could be accessed by all subclasses. This allows me to pass data from one state to another by using the mapping of the BlackBoard.
Added onStateEntry() and onStateDeparture() methods to the ScreenState class to
allow for states to perform specific behaviour at runtime when entered or
departing.

+++ SpriteSystem.h
@@ -38,8 +38,8 @@ public:
 	bool isOnScreen()
 	{
 		viewPort = window->getViewport(window->getView());
-		viewPort.left = window->getView().getCenter().x - (window->getView().getSize().x / 2);
-		viewPort.top = window->getView().getCenter().y - (window->getView().getSize().y / 2);
+		viewPort.left = (int)(window->getView().getCenter().x - (window->getView().getSize().x / 2));
+		viewPort.top = (int)(window->getView().getCenter().y - (window->getView().getSize().y / 2));
 
 		width = int(sprite->getTextureRect().width * sprite->getScale().x);
 		height = int(sprite->getTextureRect().height * sprite->getScale().y);
